make repo_config global, don't pass to scan_tree()
In cgit, rename repo_config() to cgit_repo_config() and make it global.
In scan-tree, stop accepting a repo_config_fn for all of the functions
and instead just use the new cgit_repo_config.  This makes the code
easier to read and will make these functions fit a future standard for
repository backends.

Signed-off-by: Mark Lodato <lodatom@gmail.com>