only scan when necessary; support other backends
When using the 'scan-path' option, the full recursive directory scan is
no longer done for any page but 'repolist'.  Instead, the filesystem
path is computed directly by concatenating 'scan-path' and the 'repo'
from the URL.  This offers a speedup when there are many repositories on
a slow disk.

To implement this, a generic "repolist provider" framework is
implemented.  This allows for other repository list backends to be added
in the future.

NOTE: Now, scan_tree() is called after the entire config file is read.
Previously, it was called immediately when it was encountered in the
config file.

Signed-off-by: Mark Lodato <lodatom@gmail.com>